Add Node.js 8 and 9 to the builds
Node.js 8 is an LTS release so it should be added to the CI build to
ensure that they build. Node.js 9 serves as the base for the next LTS
release, Node.js 10, so it should also be included in the CI build to
ensure that everything is fine there too.

Signed-off-by: Remy Suen <remy.suen@gmail.com>